requirements.txt
camelot-py[cv]
pandas
openpyxl
matplotlib
PyPDF2==1.26.0
tkinter